requirements.txt
Django==5.0.4
djangorestframework==3.15.1
Requests==2.31.0
